Logging commands should not take clock cycles not used.
Changed logging structure to use deferred execution, so that when I
submit a test and don't have logging ON, we are not penalized for
the time it takes to __generate__ the logging output.
This was particularly handy inside a forloop where I was logging
the contents of an array.  Even when not outputting the string, I
was still CALCULATING the value, which caused major performance
problems.
With this new method, we can capture the IDEA of the logging,
but choose not to commit to it.
It is a bit ugly to look at, though, but right now, its the best idea
I have.